+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+## [0.1.88] - 2025-10-06
+Addressing addition of duplicate application tags
+### Changed
+api/src/main/java/com/homeaway/datapullclient/process/DataPullRequestProcessor.java
+api/src/main/java/com/homeaway/datapullclient/data/ClusterStepCollection.java
 
 ## [0.1.87] - 2025-09-22
 Addressing listCluster concurrency issue by changing HashMap to ConcurrentHashMap